helm-chart-maker/pyproject.toml
[project]
name = "helm-chart-maker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama>=0.4.6",
    "email-validator>=2.2.0",
    "helm-chart-maker",
    "jinja2>=3.1.4",
    "pydantic-extra-types>=2.9.0",
    "pydantic>=2.9.1",
    "pyyaml>=6.0.2",
    "semver>=3.0.2",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
]

[tool.uv.sources]
helm-chart-maker = { workspace = true }

[tool.pytest.ini_options]
pythonpath=[
    "."
]
addopts = [
    "--import-mode=importlib",
]
